Move pyopenssl specific test code to test_openssl.
